+++ MyProfile.Entity/Model/Feedback.cs
@@ -8,7 +8,7 @@ namespace MyProfile.Entity.Model
 	{
 		[Key]
 		[DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-		public int ID { get; set; }
+		public long ID { get; set; }
 		[Required]
 		[MaxLength(32)]
 		public string Status { get; set; }
@@ -20,7 +20,7 @@ namespace MyProfile.Entity.Model
 
 
         [ForeignKey("Chat")]
-		public int ChatID { get; set; }
+		public long ChatID { get; set; }
 
 		public virtual Chat Chat { get; set; }
