+++ MyProfile.Entity/Model/Limit.cs
@@ -9,7 +9,7 @@ namespace MyProfile.Entity.Model
 	{
 		[Key]
 		[DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-		public int ID { get; set; }
+		public long ID { get; set; }
 		[Required]
 		public string Name { get; set; }
 		public string Description { get; set; }
@@ -27,7 +27,7 @@ namespace MyProfile.Entity.Model
 		[ForeignKey("PeriodType")]
 		public int PeriodTypeID { get; set; }
 		[ForeignKey("VisibleElement")]
-		public int VisibleElementID { get; set; }
+		public long VisibleElementID { get; set; }
 		[ForeignKey("Currency")]
 		public int? CurrencyID { get; set; }
