+++ MyProfile.Entity/Model/UserTag.cs
@@ -9,7 +9,7 @@ namespace MyProfile.Entity.Model
 	{
 		[Key]
 		[DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-		public int ID { get; set; }
+		public long ID { get; set; }
 		[Required]
 		[MaxLength(132)]
 		public string Title { get; set; }
@@ -22,7 +22,7 @@ namespace MyProfile.Entity.Model
 
 
         [ForeignKey("Tag")]
-		public int? TagID { get; set; }
+		public long? TagID { get; set; }
 		[ForeignKey("User")]
 		public Guid? UserID { get; set; }
